backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
faiss-cpu
PyPDF2
openai
python-dotenv
python-multipart
sentence-transformers
groq

